AuthHelper moved to separate package
